seiya-step-by-step/35_职位数和薪资Top10城市/seiya/seiya/requirements.txt
Flask
flask-login
marshmallow
hiredis
redis
Flask-SQLAlchemy
ipython
pylint
pytest
pytest-flask
flask-shell-ipython
flask-migrate
flask-wtf
mysqlclient
faker
fake-useragent
